requirements.txt
greenlet==3.3.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.46
